rename WhereAssert to WhereClauseAssert
